Refactor current commands and add skeleton file for new Commands
