+++ src/main/java/seedu/address/logic/commands/modulelesson/ClearModuleLessonCommand.java
@@ -1,0 +1,20 @@
+package seedu.address.logic.commands.modulelesson;
+
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Model;
+
+public class ClearModuleLessonCommand extends Command {
+    /**
+     * Executes the command and returns the result message.
+     *
+     * @param model {@code Model} which the command should operate on.
+     * @return feedback message of the operation result for display
+     * @throws CommandException If an error occurs during command execution.
+     */
+    @Override
+    public CommandResult execute(Model model) throws CommandException {
+        return null;
+    }
+}
