+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -11,7 +11,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TELE_HANDLE;
 import java.util.Set;
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.AddCommand;
+import seedu.address.logic.commands.person.AddCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.ModuleCode;
