+++ src/test/java/seedu/address/logic/commands/RemarkCommandTest.java
@@ -15,6 +15,8 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.person.ClearCommand;
+import seedu.address.logic.commands.person.RemarkCommand;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
